requirements.txt
pyTelegramBotAPI
ngrok-api
